+++ lib/grape/validations/validators/default_validator.rb
@@ -8,19 +8,18 @@ module Grape
           super
           # !important, lazy call at runtime
           @default_call =
-            if @options.is_a?(Proc)
-              @options.arity.zero? ? proc { @options.call } : @options
-            elsif @options.duplicable?
-              proc { @options.dup }
+            if options.is_a?(Proc)
+              options.arity.zero? ? proc { options.call } : options
+            elsif options.duplicable?
+              proc { options.dup }
             else
-              proc { @options }
+              proc { options }
             end
         end
 
         def validate!(params)
-          attrs = SingleAttributeIterator.new(@attrs, @scope, params)
-          attrs.each do |resource_params, attr_name|
-            next unless @scope.meets_dependency?(resource_params, params)
+          @iterator.each(params) do |resource_params, attr_name|
+            next unless scope.meets_dependency?(resource_params, params)
 
             resource_params[attr_name] = @default_call.call(resource_params) if hash_like?(resource_params) && resource_params[attr_name].nil?
           end